Conflicts:
	build/Android.common.mk
	compiler/dex/quick/gen_invoke.cc
	compiler/driver/compiler_driver.cc
	compiler/image_writer.cc
	runtime/Android.mk
	runtime/base/logging.cc
	runtime/base/mutex-inl.h
	runtime/base/mutex.cc
	runtime/class_linker.cc
	runtime/class_linker.h
	runtime/gc/heap.cc
	runtime/gc/space/dlmalloc_space.cc
	runtime/gc/space/dlmalloc_space.h
	runtime/gc/space/large_object_space.h
	runtime/gc/space/space.h
	runtime/hprof/hprof.cc
	runtime/indirect_reference_table_test.cc
	runtime/mirror/array-inl.h
	runtime/mirror/class-inl.h
	runtime/mirror/class.cc
	runtime/mirror/object-inl.h
	runtime/mirror/object.cc
	runtime/native/dalvik_system_DexFile.cc
	runtime/native/dalvik_system_VMRuntime.cc
	runtime/native/dalvik_system_Zygote.cc
	runtime/object_lock.h
	runtime/stack.h
	runtime/utils.cc
	runtime/verifier/method_verifier.cc
	runtime/verifier/method_verifier.h
	runtime/verifier/register_line.cc
	runtime/verifier/register_line.h

+++ compiler/output_stream.h
@@ -18,6 +18,7 @@
 #define ART_COMPILER_OUTPUT_STREAM_H_
 
 #include <stdint.h>
+#include <sys/types.h>
 
 #include <string>
 
@@ -41,7 +42,7 @@ class OutputStream {
     return location_;
   }
 
-  virtual bool WriteFully(const void* buffer, int64_t byte_count) = 0;
+  virtual bool WriteFully(const void* buffer, size_t byte_count) = 0;
 
   virtual off_t Seek(off_t offset, Whence whence) = 0;
